Update 01-welcome.html
I remember you being passionate about the unreasonable institute day one of lecture. I had a co-founder of the unreasonable institute come speak in a global development class I took. I think it is a super cool thing that needs it's own slide. It is something a lot of people may be interested in that they may miss in your misc section (going to attempt to upload the associated picture I put in the slide in the images)